Added transactions to context for now. Moved transaciton fetch and button to table for now. Moved the login transaction fetch useeffect to login for now. Made logout properly empty accounts. Adjusted spacing. Removed unused swithc code. Some temporary code for testing display.
