+++ northark-frontend/src/containers/Login.js
@@ -1,4 +1,4 @@
-import { useState, useContext } from 'react';
+import { useState, useEffect, useContext } from 'react';
 import { UserContext } from '../contexts/UserContext';
 import axios from 'axios'
 
@@ -40,8 +40,38 @@ const Login = () => {
 
   const handleLogout = () => {
     localStorage.removeItem('JWT');
-    setUserState(userState=>({...userState, loggedin: false}))
+    setUserState(userState=>({...userState, loggedin: false, accounts:[]}))
   }
+
+
+
+  useEffect (()=>{
+    const token = localStorage.getItem('JWT');
+    const fetchAccounts =  async () =>{
+      const response = await fetch('https://pristine-yosemite-12350.herokuapp.com/users/findaccountsbycustomer', {
+          headers: {Authorization: "JWT " + token},
+          },);
+      if (response.status !== 200) {
+            setUserState(userState=> ({...userState, loggedin: false, name: '', customerId:'', phone:'', email:''}));
+      } else {
+          const data = await response.json();
+          setUserState(userState=> ({...userState, 
+            loggedin: true, 
+            name: data.accounts[0].name, 
+            customerId:data.accounts[0].id, 
+            phone:data.accounts[0].phone, 
+            email:data.accounts[0].email,
+            accounts: data.accounts[0].Accounts
+          }));
+      }
+      
+      };
+      fetchAccounts()
+  }, [userState.loggedin])
+
+  
+
+
 
   if (!userState.loggedin) {
     return(
